songs-to-spotify/requirements.txt
requests
datetime
bs4

google-auth-oauthlib
google-api-python-client